Rename qr-utils to ticket-utils, switch QR to CID embedding, remove ticket card HTML
The file contained much more than QR utilities. Switched from base64 data URLs
(blocked by many email clients) to CID-attached PNGs for reliable QR rendering.
Removed buildTicketCardHtml/buildTicketListHtml in favor of CMS-managed templates.
